Release 0.3.0 cherry-pick (#602)

+++ src/sequences.h
@@ -6,7 +6,7 @@ struct Sequences {
   Sequences(std::span<const int32_t> input_sequence, int batch_size, int beam_size, int max_length);
 
   // Returns a sequence of word IDs for a given beam index ( beam_index < batch_beam_size).
-  cpu_span<int32_t> GetSequence(int batch_beam_index);
+  cpu_span<int32_t> GetSequence(size_t batch_beam_index);
   cpu_span<int32_t> GetSequences() { return sequences_; }
 
   // Returns current sequence length.
